fixed CA move piece
