+++ src/main/java/useCase/movePiece/MovePieceInputBoundary.java
@@ -3,5 +3,5 @@ package useCase.movePiece;
 import interfaceAdapters.movePiece.MovePieceViewModel;
 
 public interface MovePieceInputBoundary {
-    void movePiece(MovePieceInputData movePieceInputData, MovePieceViewModel movePieceViewModel);
+    void movePiece(MovePieceInputData movePieceInputData);
 }
